requirements.txt
pip==20.3.1
setuptools>=26
wheel>=0.29
pandas
numpy
pytest
coverage
twine
matplotlib
seaborn
gcsfs==0.7.0
simplejson
scikit-learn
statsmodels
folium
xgboost
termcolor
mlflow
memoized_property
joblib
google-cloud-storage
python-dotenv
flask==1.1.1
flask-cors
gunicorn
